+++ src/components/AppBreadcrumbs/AppBreadcrumbs.tsx
@@ -1,18 +1,6 @@
 import React from 'react';
-import { Theme, withStyles } from '@material-ui/core/styles';
-import { Breadcrumbs, Link, LinkProps } from '@material-ui/core';
-import { Link as RouterLink } from 'react-router-dom';
-import { ClassNameMap } from '@material-ui/styles';
-
-const styles = (theme: Theme) => ({
-  terminalBreadcrumb: {
-    color: theme.palette.primary.dark,
-    cursor: 'default',
-    '&:hover': {
-      textDecoration: 'none',
-    },
-  },
-});
+import { Breadcrumbs } from '@material-ui/core';
+import BreadcrumbLink from './BreadcrumbLink';
 
 type Context = { type: string; id: string; name: string };
 type Breadcrumb = { text: string; to: string };
@@ -21,24 +9,6 @@ type AppBreadcrumbsProps = {
   childBreadcrumbs: Breadcrumb[];
   context: Context;
 };
-
-const BreadcrumbLink = withStyles(styles)(
-  (props: LinkProps<RouterLink, { classes: ClassNameMap; disabled: boolean }>) => {
-    const { classes, ...other } = props;
-    if (props.disabled) {
-      return (
-        <Link
-          component={RouterLink}
-          color="primary"
-          className={classes.terminalBreadcrumb}
-          onClick={(e) => e.preventDefault()}
-          {...other}
-        />
-      );
-    }
-    return <Link component={RouterLink} color="primary" {...other} />;
-  },
-);
 
 const AppBreadcrumbs = ({ context, childBreadcrumbs }: AppBreadcrumbsProps) => {
   const { id, name, type } = context;
